+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -4,7 +4,7 @@ This page describes changes to the Airtime Service Interface implemented across 
 
 Released 25 July 2020
 
-- Added the `RetryPurchase` operation to the `PurchaseResource` resource.
+- Added the `LookupPurchase` operation to the `PurchaseResource` resource.
 - Added new fields to the `Product` class:
     - `validityPeriod`: The length of time for which a product is valid. Any unused portion of a product (e.g. unused data) will be forfeit at the end of the validity period).
     - `productContents`: A list defining the contents of the airtime product.